Update for only one active Tag
